requirements.txt
streamlit==1.28.0
tensorflow==2.13.0
opencv-python==4.8.1.78
pillow==10.0.1
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
pandas==2.1.1
numpy==1.24.3